LayoutTests/fast/files/url-null-expected.txt

Test URL methods with null arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS URL.createObjectURL(null) is null
URL.revokeObjectURL(null)
PASS successfullyParsed is true

TEST COMPLETE
